// src/lib/enums/UFCodes.enum.ts
export enum UFCodes {
  ACRE = 12,
  ALAGOAS = 27,
  BAHIA = 29,
  CEARA = 23,
  DISTRITO_FEDERAL = 53,
  ESPIRITO_SANTO = 32,
  GOIAS = 52,
  AMAPA = 16,
  AMAZONAS = 13,
  MARANHAO = 21,
  MATO_GROSSO_DO_SUL = 50,
  MATO_GROSSO = 51,
  MINAS_GERAIS = 31,
  PARA = 15,
  PARAIBA = 25,
  PARANA = 41,
  PIAUI = 22,
  RIO_DE_JANEIRO = 33,
  RIO_GRANDE_DO_NORTE = 24,
  RIO_GRANDE_DO_SUL = 43,
  RONDONIA = 11,
  RORAIMA = 14,
  SANTA_CATARINA = 42,
  PERNAMBUCO = 26,
  SERGIPE = 28,
  SAO_PAULO = 35,
  TOCANTINS = 17,
}